requirements.txt
httpx>=0.27.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
